+++ lib/hexapdf/type/page_tree_node.rb
@@ -207,9 +207,22 @@ module HexaPDF
         end
 
         parent = page[:Parent]
+        prev_index = page.index
+        prev_kid_index = parent[:Kids].index(page)
+
         insert_page(to_index, page)
+
         ancestors.each {|node| node[:Count] -= 1 }
-        parent[:Kids].delete(page)
+        if page[:Parent] == parent
+          to_index = self[:Count] + to_index + 1 if to_index < 0
+          if prev_index < to_index
+            parent[:Kids].delete_at(prev_kid_index)
+          else
+            parent[:Kids].delete_at(prev_kid_index + 1)
+          end
+        else
+          parent[:Kids].delete(page)
+        end
       end
 
       # :call-seq: